+++ frontend/package.json
@@ -126,6 +126,7 @@
     "react-absolute-grid": "^3.0.1",
     "react-awesome-popover": "^3.1.1",
     "react-click-outside": "^3.0.1",
+    "react-clipboard.js": "^2.0.7",
     "react-collapse": "^4.0.3",
     "react-color": "^2.17.1",
     "react-flip-move": "^3.0.3",
@@ -143,6 +144,7 @@
     "react-router-dom": "^4.3.1",
     "react-search-input": "^0.11.3",
     "react-svg-gauge": "^1.0.10",
+    "react-tabs": "^3.0.0",
     "react-timekeeper": "^1.1.0",
     "react-toastify": "^5.1.0",
     "react-toastify-redux": "github:omigost/react-toastify-redux",
